+++ lib/musicbrainz/models/release_group.rb
@@ -1,42 +1,53 @@
-# -*- encoding : utf-8 -*-
 module MusicBrainz
   class ReleaseGroup < BaseModel
-    field :id, String
-    field :type, String
-    field :title, String
-    field :desc, String
-    field :first_release_date, Date
-    field :urls, Hash
+    include Mapper::Resources::ReleaseGroup
+    
+    xml_accessor :mbid, from: '@id' # deprecated
+    
+    alias_method :desc, :disambiguation
+    
+    # user_tags user_ratings
+    INCLUDES = %w(artists releases artist_credits tags ratings discids media artist_credits annotation url_rels artist_rels label_rels recording_rels release_rels release_group_rels url_rels work_rels)
 
-    alias_method :disambiguation, :desc
-
-    def releases
-      @releases ||= client.load(:release, { release_group: id, inc: [:media, :release_groups], limit: 100 }, {
-        binding: :release_group_releases,
-        create_models: :release,
-        sort: :date
-      }) unless @id.nil?
+    def releases(query = {})
+      return if @id.nil?
+      
+      return @releases if @releases || do_not_search
+      
+      @releases ||= client.search(
+        'MusicBrainz::Release', { release_group: id, inc: [:media, :release_groups, :recordings], limit: 100 }.merge(query), sort: :date
+      )
     end
 
     class << self
-      def find(id)
-        client.load(:release_group, { id: id, inc: [:url_rels] }, {
-          binding: :release_group,
-          create_model: :release_group
-        })
+      def search(artist, title, options = {})
+        if artist.split('-').length != 5
+          artist = CGI.escape(artist).gsub(/\!/, '\!')  
+        end
+        
+        title = CGI.escape(title).gsub(/\!/, '\!')
+        extra_query = options[:extra_query] ? options[:extra_query] : ''
+        extra_query += " AND type:#{options[:type]}" if options[:type]
+        
+        params = {}
+        params[:offset] = options[:offset] if options.has_key? :offset
+        
+        if artist.split('-').length == 5
+          params[:query] = "arid:#{artist} AND releasegroup:\"#{title}\" #{extra_query}"
+          client.search('MusicBrainz::ReleaseGroup', params, create_models: false)
+        else
+          params[:query] = "artistname:\"#{artist}\" AND releasegroup:\"#{title}\" #{extra_query}"
+          client.search('MusicBrainz::ReleaseGroup', params, create_models: false)
+        end
       end
       
-      def search(artist_name, title, type = nil)
-				if type
-					super({artist: artist_name, releasegroup: title, type: type})
-				else
-					super({artist: artist_name, releasegroup: title})
-				end
-      end
-      
-      def find_by_artist_and_title(artist_name, title, type = nil )
-        matches = search(artist_name, title, type)
+      def find_by_artist_and_title(artist_name, title, options = {})
+        matches = search(artist_name, title, options)
         matches.nil? || matches.empty? ? nil : find(matches.first[:id])
+      end
+      
+      def find_by_artist_id(artist_id)
+        Artist.new(id: artist_id).release_groups
       end
     end
   end