+++ spec/spec_helper.rb
@@ -1,8 +1,14 @@
-# -*- encoding : utf-8 -*-
+if ENV['COVERAGE_REPORT']
+  require 'simplecov'
+  SimpleCov.start 'rails'
+end
+
 require "rubygems"
 require "bundler/setup"
 require "musicbrainz"
 require "pry"
+
+Dir[File.expand_path('./spec/support/**/*.rb')].each {|f| require f}
 
 RSpec.configure do |c|
   c.order = 'random'
@@ -18,4 +24,5 @@ MusicBrainz.configure do |c|
 
   c.cache_path = File.join(File.dirname(__FILE__), '..', 'tmp', 'spec_cache')
   c.perform_caching = true
+  c.hexdigest_url = false
 end