+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 ## musicbrainz (unreleased) ##
 
-# 0.7.7 (Dezember 1, 2014) ##
+*   Increases coverage of MusicBrainz API and replaces bindings by Music Brainz XSL schema transformation mapper.
+*   TrackSearch renamed to RecordingSearch.
+*   Find Recording by Release ID.
+*   Artist#release_groups supports :offset param.
+*   Adds :limit and :offset params to Recording search.
+*   Adds :offset option for ReleaseGroup search.
+*   Improves artist search and search everywhere and not only in artist name field.
+*   extra_query option for MusicBrainz::Artist#release_groups.
+
+# 0.7.7 (December 1, 2014) ##
 
 *   Handle nil values returned from Base model search at Artist#find_by_name and ReleaseGroup#find_by_artist_and_title [#24]